Clean up examples, provide optionality, and add tracing
